refactor: refactoring emit some edge solution
